Adjusted controller_execute to return result, activated from_result functions
